requirements.txt
streamlit==1.28.0
openai==1.35.0
azure-search-documents==11.4.0
azure-identity==1.14.0
python-dotenv==1.0.0
azure-core==1.30.0